shape code, io feature
