+++ .github/workflows/build.yml
@@ -18,7 +18,7 @@ jobs:
       - name: build
         run: cargo build
       - name: test
-        run: cargo test
+        run: cargo test --features kramer-io
         env:
           REDIS_HOST: localhost
           REDIS_PORT: ${{ job.services.redis.ports[6379] }}
