+++ src/lib.rs
@@ -4,8 +4,9 @@
 //! ## Example
 //!
 //! ```
-//! use kramer::{Command, send};
+//! use kramer::{Command, StringCommand, Arity, Insertion};
 //! use std::env::{var};
+//! use std::io::prelude::*;
 //!
 //! fn get_redis_url() -> String {
 //!   let host = var("REDIS_HOST").unwrap_or(String::from("0.0.0.0"));
@@ -16,14 +17,9 @@
 //! fn main() -> Result<(), Box<dyn std::error::Error>> {
 //!   let url = get_redis_url();
 //!   let cmd = Command::Keys("*");
-//!
-//!   async_std::task::block_on(async {
-//!     match send(url.as_str(), cmd).await {
-//!       Ok(keys) => println!("keys is a response: {:?}", keys),
-//!       Err(e) => println!("unable to send: {:?}", e),
-//!     }
-//!   });
-//!
+//!   let mut stream = std::net::TcpStream::connect(url)?;
+//!   write!(stream, "{}", cmd)?;
+//!   write!(stream, "{}", StringCommand::Set(Arity::One(("name", "kramer")), None, Insertion::Always))?;
 //!   Ok(())
 //! }
 //! ```
@@ -31,404 +27,23 @@
 //! [redis]: https://redis.io/topics/protocol
 //! [async-std]: https://github.com/async-rs/async-std
 //! [tcp-stream]: https://docs.rs/async-std/0.99.11/async_std/net/struct.TcpStream.html
-extern crate async_std;
+#[cfg(feature = "kramer-io")]
+mod io;
+#[cfg(feature = "kramer-io")]
+pub use io::{execute, send, Response, ResponseValue};
 
-use async_std::net::TcpStream;
-use async_std::prelude::*;
-use std::io::{Error as IOError, ErrorKind as IOErrorKind};
+mod modifiers;
+use modifiers::format_bulk_string;
+pub use modifiers::{Arity, Insertion, Side};
 
-#[derive(Debug)]
-enum ResponseLine {
-  Array(usize),
-  SimpleString(String),
-  Error(String),
-  Integer(i64),
-  BulkString(usize),
-  Null,
-}
+mod lists;
+pub use lists::ListCommand;
 
-#[derive(Debug, PartialEq)]
-pub enum ResponseValue {
-  Empty,
-  String(String),
-  Integer(i64),
-}
+mod strings;
+pub use strings::StringCommand;
 
-#[derive(Debug, PartialEq)]
-pub enum Response {
-  Array(Vec<ResponseValue>),
-  Item(ResponseValue),
-  Error,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Side {
-  Left,
-  Right,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Insertion {
-  Always,
-  IfExists,
-  IfNotExists,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Arity<S> {
-  Many(Vec<S>),
-  One(S),
-}
-
-#[derive(Debug)]
-pub enum ListCommand<S>
-where
-  S: std::fmt::Display,
-{
-  Len(S),
-  Push((Side, Insertion), S, Arity<S>),
-  Pop(Side, S, Option<(Option<Arity<S>>, u64)>),
-  Rem(S, S, u64),
-  Index(S, i64),
-  Set(S, u64, S),
-  Insert(S, Side, S, S),
-  Trim(S, i64, i64),
-  Range(S, i64, i64),
-}
-
-fn format_bulk_string<S: std::fmt::Display>(input: S) -> String {
-  let as_str = format!("{}", input);
-  format!("${}\r\n{}\r\n", as_str.len(), as_str)
-}
-
-impl<S: std::fmt::Display> std::fmt::Display for ListCommand<S> {
-  fn fmt(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-    match self {
-      ListCommand::Trim(key, start, stop) => {
-        let tail = format!(
-          "{}{}{}",
-          format_bulk_string(key),
-          format_bulk_string(start),
-          format_bulk_string(stop)
-        );
-        write!(formatter, "*4\r\n$5\r\nLTRIM\r\n{}", tail)
-      }
-      ListCommand::Set(key, index, element) => {
-        let tail = format!(
-          "{}{}{}",
-          format_bulk_string(key),
-          format_bulk_string(index),
-          format_bulk_string(element)
-        );
-        write!(formatter, "*4\r\n$4\r\nLSET\r\n{}", tail)
-      }
-      ListCommand::Insert(key, side, pivot, element) => {
-        let side = match side {
-          Side::Left => format_bulk_string("BEFORE"),
-          Side::Right => format_bulk_string("AFTER"),
-        };
-        let tail = format!("{}{}", format_bulk_string(pivot), format_bulk_string(element));
-
-        write!(
-          formatter,
-          "*5\r\n$7\r\nLINSERT\r\n{}{}{}",
-          format_bulk_string(key),
-          side,
-          tail,
-        )
-      }
-      ListCommand::Index(key, amt) => {
-        let tail = format!("{}{}", format_bulk_string(key), format_bulk_string(amt));
-        write!(formatter, "*3\r\n$6\r\nLINDEX\r\n{}", tail)
-      }
-      ListCommand::Rem(key, value, count) => {
-        let end = format!(
-          "{}{}{}",
-          format_bulk_string(key),
-          format_bulk_string(count),
-          format_bulk_string(value),
-        );
-
-        write!(formatter, "*4\r\n$4\r\nLREM\r\n{}", end)
-      }
-      ListCommand::Range(key, from, to) => {
-        let end = format!("{}{}", format_bulk_string(from), format_bulk_string(to));
-        write!(formatter, "*4\r\n$6\r\nLRANGE\r\n{}{}", format_bulk_string(key), end)
-      }
-      ListCommand::Len(key) => write!(formatter, "*2\r\n$4\r\nLLEN\r\n{}", format_bulk_string(key)),
-      ListCommand::Pop(side, key, block) => {
-        let (cmd, ext, kc) = match (side, block) {
-          (Side::Left, None) => ("LPOP", format!(""), 0),
-          (Side::Right, None) => ("RPOP", format!(""), 0),
-          (Side::Left, Some((None, timeout))) => ("BLPOP", format_bulk_string(timeout), 1),
-          (Side::Right, Some((None, timeout))) => ("BRPOP", format_bulk_string(timeout), 1),
-          (Side::Left, Some((Some(values), timeout))) => {
-            let (vc, ext) = match values {
-              Arity::One(value) => (1, format_bulk_string(value)),
-              Arity::Many(values) => (values.len(), values.iter().map(format_bulk_string).collect::<String>()),
-            };
-            ("BLPOP", format!("{}{}", ext, format_bulk_string(timeout)), vc + 1)
-          }
-          (Side::Right, Some((Some(values), timeout))) => {
-            let (vc, ext) = match values {
-              Arity::One(value) => (1, format_bulk_string(value)),
-              Arity::Many(values) => (values.len(), values.iter().map(format_bulk_string).collect::<String>()),
-            };
-            ("BRPOP", format!("{}{}", ext, format_bulk_string(timeout)), vc + 1)
-          }
-        };
-        write!(
-          formatter,
-          "*{}\r\n${}\r\n{}\r\n{}{}",
-          2 + kc,
-          cmd.len(),
-          cmd,
-          format_bulk_string(key),
-          ext
-        )
-      }
-      ListCommand::Push(operation, k, Arity::One(v)) => {
-        let cmd = match operation {
-          (Side::Left, Insertion::IfExists) => "LPUSHX",
-          (Side::Right, Insertion::IfExists) => "RPUSHX",
-          (Side::Left, _) => "LPUSH",
-          (Side::Right, _) => "RPUSH",
-        };
-        let parts = format!("{}{}", format_bulk_string(k), format_bulk_string(v),);
-        write!(formatter, "*3\r\n${}\r\n{}\r\n{}", cmd.len(), cmd, parts)
-      }
-      ListCommand::Push(operation, k, Arity::Many(v)) => {
-        let size = v.len();
-        let cmd = match operation {
-          (Side::Left, Insertion::IfExists) => "LPUSHX",
-          (Side::Right, Insertion::IfExists) => "RPUSHX",
-          (Side::Left, _) => "LPUSH",
-          (Side::Right, _) => "RPUSH",
-        };
-        let parts = format!(
-          "{}{}",
-          format_bulk_string(k),
-          v.iter().map(format_bulk_string).collect::<String>()
-        );
-        write!(formatter, "*{}\r\n${}\r\n{}\r\n{}", 2 + size, cmd.len(), cmd, parts)
-      }
-    }
-  }
-}
-
-#[derive(Debug)]
-pub enum StringCommand<S>
-where
-  S: std::fmt::Display,
-{
-  Set(Arity<(S, S)>, Option<std::time::Duration>, Insertion),
-  Get(Arity<S>),
-  Decr(S, usize),
-  Incr(S, i64),
-  Append(S, S),
-}
-
-impl<S: std::fmt::Display> std::fmt::Display for StringCommand<S> {
-  fn fmt(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-    match self {
-      StringCommand::Incr(key, 1) => write!(formatter, "*2\r\n$4\r\nINCR\r\n{}", format_bulk_string(key)),
-      StringCommand::Incr(key, amt) => write!(
-        formatter,
-        "*3\r\n$6\r\nINCRBY\r\n{}{}",
-        format_bulk_string(key),
-        format_bulk_string(amt)
-      ),
-      StringCommand::Decr(key, 1) => write!(formatter, "*2\r\n$4\r\nDECR\r\n{}", format_bulk_string(key)),
-      StringCommand::Decr(key, amt) => write!(
-        formatter,
-        "*3\r\n$6\r\nDECRBY\r\n{}{}",
-        format_bulk_string(key),
-        format_bulk_string(amt)
-      ),
-      StringCommand::Get(Arity::One(key)) => write!(formatter, "*2\r\n$3\r\nGET\r\n{}", format_bulk_string(key)),
-      StringCommand::Get(Arity::Many(keys)) => {
-        let count = keys.len();
-        let tail = keys.iter().map(format_bulk_string).collect::<String>();
-        write!(formatter, "*{}\r\n$4\r\nMGET\r\n{}", count + 1, tail)
-      }
-      StringCommand::Append(key, value) => write!(
-        formatter,
-        "*3\r\n$6\r\nAPPEND\r\n{}{}",
-        format_bulk_string(key),
-        format_bulk_string(value)
-      ),
-      StringCommand::Set(Arity::One((key, value)), timeout, insertion) => {
-        let (k, v) = (format_bulk_string(key), format_bulk_string(value));
-        let (cx, px) = match timeout {
-          None => (0, format!("")),
-          Some(t) => (
-            2,
-            format!("{}{}", format_bulk_string("PX"), format_bulk_string(t.as_millis())),
-          ),
-        };
-        let (ci, i) = match insertion {
-          Insertion::IfExists => (1, format_bulk_string("XX")),
-          Insertion::IfNotExists => (1, format_bulk_string("NX")),
-          Insertion::Always => (0, format!("")),
-        };
-        write!(formatter, "*{}\r\n$3\r\nSET\r\n{}{}{}{}", 3 + ci + cx, k, v, px, i)
-      }
-      // Timeouts are not supported with a many set.
-      StringCommand::Set(Arity::Many(assignments), _, insertion) => {
-        let count = (assignments.len() * 2) + 1;
-        let cmd = match insertion {
-          Insertion::IfNotExists => "MSETNX",
-          _ => "MSET",
-        };
-        let tail = assignments
-          .iter()
-          .map(|(k, v)| format!("{}{}", format_bulk_string(k), format_bulk_string(v)))
-          .collect::<String>();
-        write!(formatter, "*{}\r\n{}{}", count, format_bulk_string(cmd), tail)
-      }
-    }
-  }
-}
-
-#[derive(Debug)]
-pub enum HashCommand<S>
-where
-  S: std::fmt::Display,
-{
-  Del(S, S, Option<Arity<S>>),
-  Set(S, Arity<(S, S)>, Insertion),
-  Get(S, Option<Arity<S>>),
-  StrLen(S, S),
-  Len(S),
-  Incr(S, S, i64),
-  Keys(S),
-  Vals(S),
-  Exists(S, S),
-}
-
-impl<S: std::fmt::Display> std::fmt::Display for HashCommand<S> {
-  fn fmt(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-    match self {
-      HashCommand::StrLen(key, field) => {
-        let tail = format!("{}{}", format_bulk_string(key), format_bulk_string(field));
-        write!(formatter, "*3\r\n$7\r\nHSTRLEN\r\n{}", tail)
-      }
-      HashCommand::Incr(key, field, amt) => {
-        let tail = format!(
-          "{}{}{}",
-          format_bulk_string(key),
-          format_bulk_string(field),
-          format_bulk_string(amt)
-        );
-        write!(formatter, "*4\r\n$7\r\nHINCRBY\r\n{}", tail)
-      }
-      HashCommand::Vals(key) => write!(formatter, "*2\r\n$5\r\nHVALS\r\n{}", format_bulk_string(key)),
-      HashCommand::Keys(key) => write!(formatter, "*2\r\n$5\r\nHKEYS\r\n{}", format_bulk_string(key)),
-      HashCommand::Len(key) => write!(formatter, "*2\r\n$4\r\nHLEN\r\n{}", format_bulk_string(key)),
-      HashCommand::Get(key, None) => write!(formatter, "*2\r\n$7\r\nHGETALL\r\n{}", format_bulk_string(key)),
-      HashCommand::Get(key, Some(Arity::One(field))) => write!(
-        formatter,
-        "*3\r\n$4\r\nHGET\r\n{}{}",
-        format_bulk_string(key),
-        format_bulk_string(field)
-      ),
-      HashCommand::Get(key, Some(Arity::Many(fields))) => {
-        let len = fields.len();
-
-        // Awkward; Get("foo", Some(Arity::Many(vec![]))) == Get("foo", None)
-        if len == 0 {
-          let formatted = format!("{}", key);
-          return write!(formatter, "{}", HashCommand::Get(formatted, None));
-        }
-
-        let tail = fields.iter().map(format_bulk_string).collect::<String>();
-
-        write!(
-          formatter,
-          "*{}\r\n$5\r\nHMGET\r\n{}{}",
-          2 + len,
-          format_bulk_string(key),
-          tail
-        )
-      }
-      HashCommand::Exists(key, field) => write!(
-        formatter,
-        "*3\r\n$7\r\nHEXISTS\r\n{}{}",
-        format_bulk_string(key),
-        format_bulk_string(field)
-      ),
-      HashCommand::Set(key, Arity::One((field, value)), Insertion::IfNotExists) => write!(
-        formatter,
-        "*4\r\n$6\r\nHSETNX\r\n{}{}{}",
-        format_bulk_string(key),
-        format_bulk_string(field),
-        format_bulk_string(value)
-      ),
-      HashCommand::Set(key, Arity::Many(mappings), Insertion::IfNotExists) => {
-        let count = mappings.len();
-        let tail = mappings
-          .iter()
-          .map(|(k, v)| format!("{}{}", format_bulk_string(k), format_bulk_string(v)))
-          .collect::<String>();
-
-        write!(
-          formatter,
-          "*{}\r\n$6\r\nHSETNX\r\n{}{}",
-          2 + (count * 2),
-          format_bulk_string(key),
-          tail
-        )
-      }
-      HashCommand::Set(key, Arity::One((field, value)), _) => write!(
-        formatter,
-        "*4\r\n$4\r\nHSET\r\n{}{}{}",
-        format_bulk_string(key),
-        format_bulk_string(field),
-        format_bulk_string(value)
-      ),
-      HashCommand::Set(key, Arity::Many(mappings), _) => {
-        let count = mappings.len();
-        let tail = mappings
-          .iter()
-          .map(|(k, v)| format!("{}{}", format_bulk_string(k), format_bulk_string(v)))
-          .collect::<String>();
-
-        write!(
-          formatter,
-          "*{}\r\n$4\r\nHSET\r\n{}{}",
-          2 + (count * 2),
-          format_bulk_string(key),
-          tail
-        )
-      }
-      HashCommand::Del(key, field, None) => write!(
-        formatter,
-        "*3\r\n$4\r\nHDEL\r\n{}{}",
-        format_bulk_string(key),
-        format_bulk_string(field)
-      ),
-      HashCommand::Del(key, field, Some(Arity::One(s))) => write!(
-        formatter,
-        "*4\r\n$4\r\nHDEL\r\n{}{}{}",
-        format_bulk_string(key),
-        format_bulk_string(field),
-        format_bulk_string(s)
-      ),
-      HashCommand::Del(key, field, Some(Arity::Many(s))) => {
-        let count = s.len();
-        let bits = s.iter().map(format_bulk_string).collect::<String>();
-        write!(
-          formatter,
-          "*{}\r\n$4\r\nHDEL\r\n{}{}{}",
-          count + 3,
-          format_bulk_string(key),
-          format_bulk_string(field),
-          bits
-        )
-      }
-    }
-  }
-}
+mod hashes;
+pub use hashes::HashCommand;
 
 #[derive(Debug)]
 pub enum Command<S>
@@ -466,116 +81,6 @@ impl<S: std::fmt::Display> std::fmt::Display for Command<S> {
       Command::Hashes(hash_command) => write!(formatter, "{}", hash_command),
     }
   }
-}
-
-fn read_line_size(line: String) -> Result<Option<usize>, IOError> {
-  match line.split_at(1).1 {
-    "-1" => Ok(None),
-    value => value
-      .parse::<usize>()
-      .map_err(|e| {
-        IOError::new(
-          IOErrorKind::Other,
-          format!("invalid array length value '{}': {}", line.as_str(), e),
-        )
-      })
-      .map(|v| Some(v)),
-  }
-}
-
-fn readline(result: Option<Result<String, IOError>>) -> Result<ResponseLine, IOError> {
-  let line = result.ok_or_else(|| IOError::new(IOErrorKind::Other, "no line to work with"))??;
-
-  match line.bytes().next() {
-    Some(b'*') => match read_line_size(line)? {
-      None => Ok(ResponseLine::Null),
-      Some(size) => Ok(ResponseLine::Array(size)),
-    },
-    Some(b'$') => match read_line_size(line)? {
-      Some(size) => Ok(ResponseLine::BulkString(size)),
-      None => Ok(ResponseLine::Null),
-    },
-    Some(b'-') => Ok(ResponseLine::Error(line)),
-    Some(b'+') => Ok(ResponseLine::SimpleString(String::from(line.split_at(1).1))),
-    Some(b':') => {
-      let (_, rest) = line.split_at(1);
-      rest
-        .parse::<i64>()
-        .map_err(|e| IOError::new(IOErrorKind::Other, format!("{:?}", e)))
-        .and_then(|v| Ok(ResponseLine::Integer(v)))
-    }
-    Some(unknown) => Err(IOError::new(
-      IOErrorKind::Other,
-      format!("invalid message byte leader: {}", unknown),
-    )),
-    None => Err(IOError::new(
-      IOErrorKind::Other,
-      "empty line in response, unable to determine type",
-    )),
-  }
-}
-
-pub async fn execute<C, S>(mut connection: C, message: Command<S>) -> Result<Response, IOError>
-where
-  S: std::fmt::Display,
-  C: async_std::io::Write + async_std::io::Read + std::marker::Unpin,
-{
-  write!(connection, "{}", message).await?;
-
-  let mut lines = async_std::io::BufReader::new(connection).lines();
-
-  match readline(lines.next().await) {
-    Ok(ResponseLine::Array(size)) => {
-      let mut store = Vec::with_capacity(size);
-
-      if size == 0 {
-        return Ok(Response::Array(vec![]));
-      }
-
-      while let Ok(kind) = readline(lines.next().await) {
-        match kind {
-          ResponseLine::BulkString(size) => match lines.next().await {
-            Some(Ok(bulky)) if bulky.len() == size => {
-              store.push(ResponseValue::String(bulky));
-            }
-            _ => break,
-          },
-          _ => break,
-        }
-
-        if store.len() >= size {
-          return Ok(Response::Array(store));
-        }
-      }
-
-      Ok(Response::Array(store))
-    }
-    Ok(ResponseLine::BulkString(size)) => {
-      if size < 1 {
-        return Ok(Response::Item(ResponseValue::Empty));
-      }
-
-      let out = lines
-        .next()
-        .await
-        .ok_or_else(|| IOError::new(IOErrorKind::Other, "no line to work with"))??;
-
-      Ok(Response::Item(ResponseValue::String(out)))
-    }
-    Ok(ResponseLine::Null) => Ok(Response::Item(ResponseValue::Empty)),
-    Ok(ResponseLine::SimpleString(simple)) => Ok(Response::Item(ResponseValue::String(simple))),
-    Ok(ResponseLine::Integer(value)) => Ok(Response::Item(ResponseValue::Integer(value))),
-    Ok(ResponseLine::Error(e)) => Err(IOError::new(IOErrorKind::Other, e)),
-    Err(e) => Err(e),
-  }
-}
-
-pub async fn send<S>(addr: &str, message: Command<S>) -> Result<Response, IOError>
-where
-  S: std::fmt::Display,
-{
-  let mut stream = TcpStream::connect(addr).await?;
-  execute(&mut stream, message).await
 }
 
 #[cfg(test)]
@@ -901,7 +406,7 @@ mod fmt_tests {
 
   #[test]
   fn test_hdel_single() {
-    let cmd = Command::Hashes(HashCommand::Del("seinfeld", "kramer", None));
+    let cmd = Command::Hashes(HashCommand::Del("seinfeld", Arity::One("kramer")));
     let mut buffer = Vec::new();
     write!(buffer, "{}", cmd).expect("was able to write");
     assert_eq!(
@@ -912,7 +417,7 @@ mod fmt_tests {
 
   #[test]
   fn test_hdel_many() {
-    let cmd = Command::Hashes(HashCommand::Del("seinfeld", "kramer", Some(Arity::Many(vec!["jerry"]))));
+    let cmd = Command::Hashes(HashCommand::Del("seinfeld", Arity::Many(vec!["kramer", "jerry"])));
     let mut buffer = Vec::new();
     write!(buffer, "{}", cmd).expect("was able to write");
     assert_eq!(
